Added deprecated versions of the moved files, that export the new. Changed all the old imports, updated changelog
