+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/EpochBoundary.hs
@@ -1,0 +1,7 @@
+module Cardano.Ledger.Shelley.EpochBoundary
+  {-# DEPRECATED "Use 'import Cardano.Ledger.EpochBoundary' instead" #-}
+  ( module X,
+  )
+where
+
+import Cardano.Ledger.EpochBoundary as X
