+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/PoolParams.hs
@@ -1,0 +1,7 @@
+module Cardano.Ledger.Shelley.PoolParams
+  {-# DEPRECATED "Use 'import Cardano.Ledger.PoolParams' instead" #-}
+  ( module X,
+  )
+where
+
+import Cardano.Ledger.PoolParams as X
