+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rewards.hs
@@ -45,9 +45,9 @@ import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Core (EraCrypto, PParams, Reward (..), RewardType (..))
 import Cardano.Ledger.Credential (Credential (..))
 import qualified Cardano.Ledger.Crypto as CC (Crypto)
+import Cardano.Ledger.EpochBoundary (Stake (..), maxPool')
 import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
 import Cardano.Ledger.Shelley.Delegation.PoolParams (poolSpec)
-import Cardano.Ledger.Shelley.EpochBoundary (Stake (..), maxPool')
 import qualified Cardano.Ledger.Shelley.HardForks as HardForks
 import Cardano.Ledger.Shelley.TxBody (PoolParams (..))
 import Cardano.Ledger.Val ((<->))
