+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Snap.hs
@@ -22,14 +22,14 @@ import Cardano.Ledger.Coin (Coin, CompactForm)
 import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Core (EraTxOut)
 import Cardano.Ledger.Credential (Credential)
-import Cardano.Ledger.Era (EraCrypto)
-import Cardano.Ledger.Keys (KeyHash, KeyRole (StakePool, Staking))
-import Cardano.Ledger.Shelley.EpochBoundary
+import Cardano.Ledger.EpochBoundary
   ( SnapShot (_delegations, _stake),
     SnapShots (_feeSS, _pstakeGo, _pstakeMark, _pstakeSet),
     Stake (unStake),
     emptySnapShots,
   )
+import Cardano.Ledger.Era (EraCrypto)
+import Cardano.Ledger.Keys (KeyHash, KeyRole (StakePool, Staking))
 import Cardano.Ledger.Shelley.Era (ShelleySNAP)
 import Cardano.Ledger.Shelley.LedgerState
   ( DPState (..),
